--- agents-webapi-py/src/BUILD
@@ -1,14 +1,0 @@
-load("@3rdparty//:requirements.bzl", "requirement")
-package(default_visibility = ["//visibility:public"])
-
-py_binary(
-    name = "agents-webapi-app",
-    srcs = ["agents-webapi-app.py"],
-    deps = [
-        requirement("flask"),
-        requirement("jinja2"),
-        requirement("werkzeug"),
-        requirement("itsdangerous"),
-        requirement("click"),
-    ],
-)
