+++ agents_webapi_py/src/agents_webapi_app.py
@@ -1,6 +1,8 @@
 from flask import Flask
 from flask import request
 import logging
+import os
+dir_path = os.path.dirname(os.path.realpath(__file__))
 app = Flask(__name__)
 
 
@@ -14,11 +16,13 @@ def Welcome():
 def commands():
     commands = request.get_json()
     command_id = commands['id']
-    argument_count = Int(request_data['argument_count'])
+    print("command_id: {}".format(command_id))
+    argument_count = int(commands['argument_count'])
     arguments = commands['arguments']
-    for index in xrange(argument_count):
-        command_name = arguments[index]['name']
+    for index in range(argument_count):
+        argument_name = arguments[index]['name']
         command_value = arguments[index]['value']
+        print("name:  {} value: {}".format(argument_name, command_value))
     #executing command on agents-middleware
     return {'executed': True}
