+++ agents_webapi_py/src/readme.md
@@ -1,6 +1,6 @@
 ##WebApi server - running the server
 
-bazel run //agents-webapi-py/src:agents-webapi_app
+bazel run //agents-webapi-py/src:agents-webapi-app
 
 ##WebApi server - running the server tests
 bazel test //agents-webapi-py/test:agents-webapi-app-tests
